requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
azure-cognitiveservices-speech==1.25.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
ffmpeg-python==0.2.0
filelock==3.9.0
frozenlist==1.3.3
future==0.18.3
huggingface-hub==0.12.0
idna==3.4
more-itertools==9.0.0
multidict==6.0.4
numpy==1.24.1
openai==0.26.4
openai-whisper==20230124
packaging==23.0
PyAudio==0.2.13
pydub==0.25.1
python-dotenv==0.21.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
speech-synthesis==1.4
SpeechRecognition==3.9.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.0
typing_extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2
